Refactor: Clean up code formatting
