+++ src/main/java/com/activecourses/upwork/controller/user/UserController.java
@@ -3,7 +3,7 @@ package com.activecourses.upwork.controller.user;
 import com.activecourses.upwork.dto.ResponseDto;
 import com.activecourses.upwork.dto.user.UserProfileDto;
 import com.activecourses.upwork.dto.user.UserResponseDto;
-import com.activecourses.upwork.service.user.UserProfileService;
+import com.activecourses.upwork.service.user.profile.UserProfileService;
 import com.activecourses.upwork.service.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -77,7 +77,7 @@ public class UserController {
             @PathVariable int userId,
             @RequestBody @Valid UserProfileDto updateRequest) {
         try {
-            return userProfileService.UpdateUserProfile(userId,updateRequest);
+            return userProfileService.UpdateUserProfile(userId, updateRequest);
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(ResponseDto.builder()
                     .status(HttpStatus.BAD_REQUEST)
