+++ src/main/java/com/activecourses/upwork/service/user/UserServiceImpl.java
@@ -23,7 +23,6 @@ public class UserServiceImpl implements UserService {
     private final UserDtoMapper userDtoMapper;
 
 
-
     @Override
     public UserResponseDto getAllUsers(int pageNo, int pageSize, String sortBy, String sortDir) {
         // Set the sorting direction
@@ -54,11 +53,10 @@ public class UserServiceImpl implements UserService {
         userResponseDto.setLast(pagedResult.isLast());
         return userResponseDto;
     }
+
     @Override
     public User findByEmail(String email) {
         return userRepository.findByEmail(email).
-                orElseThrow(()->new UsernameNotFoundException("User Not Found"));
+                orElseThrow(() -> new UsernameNotFoundException("User Not Found"));
     }
-
-
 }
