+++ src/main/java/com/activecourses/upwork/service/user/profile/ImplUserProfileService.java
@@ -1,4 +1,4 @@
-package com.activecourses.upwork.service.user;
+package com.activecourses.upwork.service.user.profile;
 
 import com.activecourses.upwork.dto.ResponseDto;
 import com.activecourses.upwork.dto.user.UserProfileDto;
@@ -14,9 +14,10 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ImplUserProfileService implements UserProfileService{
+public class ImplUserProfileService implements UserProfileService {
     private final UserRepository userRepository;
     private final Mapper<User, UserProfileDto> userProfileMapper;
+
     @Override
     public ResponseEntity<?> getUserProfile(int id) {
         User user = userRepository.findById(id).
@@ -31,7 +32,7 @@ public class ImplUserProfileService implements UserProfileService{
                         .build());
     }
 
-    public UserProfile createUserProfile(User user){
+    public UserProfile createUserProfile(User user) {
         return UserProfile.builder()
                 .user(user).build();
     }
@@ -40,34 +41,34 @@ public class ImplUserProfileService implements UserProfileService{
     public ResponseEntity<?> UpdateUserProfile(int userId, UserProfileDto updateRequest) {
         User user = userRepository.findById(updateRequest.getId()).
                 orElseThrow(() -> new UsernameNotFoundException("user Not Found"));
-        // check if this user try to update another user profile
-        if(userId!=user.getId()){
+        // Check if this user try to update another user profile
+        if (userId != user.getId()) {
             throw new RuntimeException("this User can't update another user profile");
         }
         return ResponseEntity.ok()
                 .body(ResponseDto.builder()
                         .status(HttpStatus.OK)
                         .success(true)
-                        .data(doUpdate(updateRequest,user))
+                        .data(doUpdate(updateRequest, user))
                         .build());
     }
 
     private UserProfileDto doUpdate(UserProfileDto updateRequest, User user) {
-            //update fields that in user table
-            user.setFirstName(updateRequest.getFirstName());
-            user.setLastName(updateRequest.getLastName());
+        // Update fields that in user table
+        user.setFirstName(updateRequest.getFirstName());
+        user.setLastName(updateRequest.getLastName());
 
-            //get userProfile and update it.
-            UserProfile userProfile=user.getUserProfile();
-            userProfile.setDescription(updateRequest.getDescription());
-            userProfile.setLocation(updateRequest.getLocation());
-            userProfile.setTitle(userProfile.getTitle());
-            userProfile.setHourlyRate(updateRequest.getHourlyRate());
+        // Get userProfile and update it.
+        UserProfile userProfile = user.getUserProfile();
+        userProfile.setDescription(updateRequest.getDescription());
+        userProfile.setLocation(updateRequest.getLocation());
+        userProfile.setTitle(userProfile.getTitle());
+        userProfile.setHourlyRate(updateRequest.getHourlyRate());
 
-            // update user in database
-            user.setUserProfile(userProfile);
-            userRepository.save(user);
-            return updateRequest;
+        // Update user in database
+        user.setUserProfile(userProfile);
+        userRepository.save(user);
+        return updateRequest;
     }
 
 }
