+++ src/main/java/com/activecourses/upwork/service/user/profile/UserProfileService.java
@@ -1,4 +1,4 @@
-package com.activecourses.upwork.service.user;
+package com.activecourses.upwork.service.user.profile;
 
 import com.activecourses.upwork.dto.user.UserProfileDto;
 import com.activecourses.upwork.model.User;
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 
 public interface UserProfileService {
     ResponseEntity<?> getUserProfile(int id);
+
     UserProfile createUserProfile(User user);
 
     ResponseEntity<?> UpdateUserProfile(int userId, UserProfileDto updateRequest);
